requirements.txt
alembic==1.0.7
apiclient==1.0.3
flask==1.0.2
flask_sqlalchemy==2.3.1
flask_migrate==2.3.1
flask_script==2.0.6
google_auth_oauthlib==0.2.0
google-api-python-client==1.7.8
gunicorn==19.9.0
httplib2==0.12.0
oauth2client==4.1.3
psycopg2-binary==2.7.6.1
requests==2.21.0
